requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
ffmpeg-python==0.2.0
python-magic==0.4.27
aiofiles==23.2.1
jinja2==3.1.2